feat: reorganize sidebar {}
