+++ next-docs/pages/components/typography/private/InternalText.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const InternalText: React.FC<{}> = ({ children }) => (
+  <p className="text-lg mt-4">{children}</p>
+);
+
+export default InternalText;
